+++ Assets/Abilities/Spells.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 09080f98691131847ad856edfd8974ff
+guid: dbf6b8537431fb14792aedb66953331a
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
