+++ Assets/Scenes/AStar.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.18028328, g: 0.22571322, b: 0.30692163, a: 1}
+  m_IndirectSpecularColor: {r: 0.18028378, g: 0.22571412, b: 0.30692285, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -6248,6 +6248,10 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1511487489}
     m_Modifications:
+    - target: {fileID: 121134, guid: 1018a369bf1e2304693a74ea94205090, type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 440136, guid: 1018a369bf1e2304693a74ea94205090, type: 3}
       propertyPath: m_LocalPosition.x
       value: 35.36
@@ -9276,6 +9280,10 @@ PrefabInstance:
       propertyPath: m_Name
       value: SM_Bld_House_Chimney_04
       objectReference: {fileID: 0}
+    - target: {fileID: 178018, guid: 52d2ac88e4266ae4287982747ca76a1c, type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 493852, guid: 52d2ac88e4266ae4287982747ca76a1c, type: 3}
       propertyPath: m_LocalPosition.x
       value: 37.579998
@@ -10184,6 +10192,10 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1511487489}
     m_Modifications:
+    - target: {fileID: 188096, guid: 99b22ade5e5fef24b82c5cba6e333937, type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 422898, guid: 99b22ade5e5fef24b82c5cba6e333937, type: 3}
       propertyPath: m_LocalPosition.x
       value: 35.596
@@ -15602,6 +15614,11 @@ PrefabInstance:
       propertyPath: field of view
       value: 60
       objectReference: {fileID: 0}
+    - target: {fileID: 7502461644539187352, guid: 71f25190d1f71214e8bfb5211e776821,
+        type: 3}
+      propertyPath: m_RendererIndex
+      value: -1
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 71f25190d1f71214e8bfb5211e776821, type: 3}
 --- !u!1001 &635949752
@@ -20213,6 +20230,10 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1511487489}
     m_Modifications:
+    - target: {fileID: 177110, guid: 27e4689e1c1d6f84da2be4e3b00213dd, type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 465068, guid: 27e4689e1c1d6f84da2be4e3b00213dd, type: 3}
       propertyPath: m_LocalPosition.x
       value: 37.02
@@ -26253,6 +26274,10 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1511487489}
     m_Modifications:
+    - target: {fileID: 188096, guid: 99b22ade5e5fef24b82c5cba6e333937, type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 422898, guid: 99b22ade5e5fef24b82c5cba6e333937, type: 3}
       propertyPath: m_LocalPosition.x
       value: 46.177998
@@ -29260,6 +29285,10 @@ PrefabInstance:
       propertyPath: m_Name
       value: SM_Bld_House_Window_04
       objectReference: {fileID: 0}
+    - target: {fileID: 188096, guid: 99b22ade5e5fef24b82c5cba6e333937, type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 422898, guid: 99b22ade5e5fef24b82c5cba6e333937, type: 3}
       propertyPath: m_LocalPosition.x
       value: 43.488
@@ -29570,6 +29599,10 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1511487489}
     m_Modifications:
+    - target: {fileID: 182140, guid: f101780f32b5af64bb3c14fbf09b1771, type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 403480, guid: f101780f32b5af64bb3c14fbf09b1771, type: 3}
       propertyPath: m_LocalPosition.x
       value: 38.261
@@ -33107,6 +33140,10 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1511487489}
     m_Modifications:
+    - target: {fileID: 133646, guid: 138f01e248f319c4388a7199ae4e3eae, type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 411084, guid: 138f01e248f319c4388a7199ae4e3eae, type: 3}
       propertyPath: m_LocalPosition.x
       value: 37.02
@@ -36558,6 +36595,10 @@ PrefabInstance:
     - target: {fileID: 188096, guid: 99b22ade5e5fef24b82c5cba6e333937, type: 3}
       propertyPath: m_Name
       value: SM_Bld_House_Window_04
+      objectReference: {fileID: 0}
+    - target: {fileID: 188096, guid: 99b22ade5e5fef24b82c5cba6e333937, type: 3}
+      propertyPath: m_IsActive
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 422898, guid: 99b22ade5e5fef24b82c5cba6e333937, type: 3}
       propertyPath: m_LocalPosition.x
@@ -56432,7 +56473,7 @@ GameObject:
   - component: {fileID: 5165644636097227996}
   - component: {fileID: 5165644636097227997}
   m_Layer: 0
-  m_Name: Enemy Manager
+  m_Name: EnemyFactory
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -56464,8 +56505,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e548576c009654440ae167fd8d1546cf, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  stageDifficulty: 5
-  numberOfEnemies: 5
+  stageDifficulty: 1
+  numberOfEnemies: 2
   playerPrefab: {fileID: 3159141201095183994, guid: ef722bf4f72be6c46b0c7f6acd355a31,
     type: 3}
   camManager: {fileID: 5165644637623274078}
