+++ Assets/Scenes/Town Map/LightingData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 836e11734cd59544e88056de9acb8ddf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
