+++ Assets/Scripts/Battlefield/AStar.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 05dd861ae113a5b4fa2247d681686721
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
