--- Assets/Scripts/Battlefield/TurnBrain.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aeaa5aebf52949845bb2d288f2ae5e12
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
